greentea-bot/greentea-bot/requirements.txt
aiohttp==3.6.3
colorama==0.4.4
discord.py==1.5.1
emoji==0.6.0